+++ frontend/src/app/_services/mobile-menu-helper.service.ts
@@ -1,0 +1,19 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class MobileMenuHelperService {
+
+  private menuDisplayed = false;
+
+  constructor() { }
+
+  get isMobileMenuDisplayed() {
+    return this.menuDisplayed;
+  }
+
+  toggleMobileMenu() {
+    this.menuDisplayed = !this.menuDisplayed;
+  }
+}