+++ documentation/Changelog.md
@@ -8,7 +8,12 @@
 }
 ```
 
-### [2020-05][1.2.3]  
+### NEXT
+ 
+Bugs fixed :
+- [705 - Relative links to children directories do not work](https://github.com/KelkooGroup/theGardener/issues/705)
+
+### [2020-05][1.2.4]  
 
 The main feature is:
 - [0568 - Include OpenAPI endpoints](https://github.com/KelkooGroup/theGardener/issues/568)